Set the repository source to download the asset
When using the tool `gh` in a action, this tool can identify
the repository in the current working directory.
In the use case of an action, we are in the repository
that calls the action (`Cloudserver`, `Vault`, ...).
We want to download the asset from the repository
`action-prom-render-test` so must set this source explicitly.